Supplier to Manufacturer

+++ src/main/java/com/sales/crm/dao/AreaDAOImpl.java
@@ -3,6 +3,7 @@ package com.sales.crm.dao;
 import java.util.ArrayList;
 import java.util.Date;
 import java.util.List;
+import java.util.UUID;
 
 import org.apache.log4j.Logger;
 import org.hibernate.Query;
@@ -15,6 +16,7 @@ import org.springframework.stereotype.Repository;
 
 import com.sales.crm.model.Area;
 import com.sales.crm.model.Beat;
+import com.sales.crm.model.EntityStatusEnum;
 
 @Repository("areaDAO")
 public class AreaDAOImpl implements AreaDAO {
@@ -33,6 +35,8 @@ public class AreaDAOImpl implements AreaDAO {
 			session = sessionFactory.openSession();
 			transaction = session.beginTransaction();
 			area.setDateCreated(new Date());
+			area.setCode(UUID.randomUUID().toString());
+			area.setStatusID(EntityStatusEnum.ACTIVE.getEntityStatus());
 			session.save(area);
 			transaction.commit();
 		}catch(Exception e){
@@ -48,29 +52,37 @@ public class AreaDAOImpl implements AreaDAO {
 		}
 	}
 
+	
+	private Beat getBeatForArea(Session session, int areaID, int tenantID) {
+		SQLQuery  beatQuery = session.createSQLQuery("SELECT a.* FROM BEAT a, BEAT_AREA b WHERE a.ID = b.BEAT_ID and a.TENANT_ID = b.TENANT_ID and b.AREA_ID= ? and a.TENANT_ID = ?");
+		beatQuery.setParameter(0, areaID);
+		beatQuery.setParameter(1, tenantID);
+		List beatList = beatQuery.list();
+		Beat beat = null;
+		if(beatList != null && beatList.size() > 0){
+			for (Object obj : beatList) {
+				Object[] objs = (Object[]) obj;
+				beat = new Beat();
+				beat.setBeatID(Integer.valueOf(String.valueOf(objs[0])));
+				beat.setCode(String.valueOf(objs[1]));
+				beat.setName(String.valueOf(objs[2]));
+				beat.setDescription(String.valueOf(objs[3]));
+				beat.setCoverageSchedule(String.valueOf(objs[4]));
+				beat.setDistance(Integer.valueOf(String.valueOf(objs[5])));
+				beat.setTenantID(Integer.valueOf(String.valueOf(objs[6])));
+			}
+		}
+		return beat;
+	}
 	@Override
-	public Area get(int areaID) {
+	public Area get(int areaID, int tenantID) {
 		Session session = null;
 		Area area = null;
 		Beat beat = null;
 		try{
 			session = sessionFactory.openSession();
 			area = (Area)session.get(Area.class, areaID);
-			SQLQuery  beatQuery = session.createSQLQuery("SELECT a.* FROM BEAT a, BEAT_AREA b WHERE a.ID = b.BEAT_ID and b.AREA_ID= ?");
-			beatQuery.setParameter(0, areaID);
-			List beatList = beatQuery.list();
-			if(beatList != null && beatList.size() > 0){
-				for (Object obj : beatList) {
-					Object[] objs = (Object[]) obj;
-					beat = new Beat();
-					beat.setBeatID(Integer.valueOf(String.valueOf(objs[0])));
-					beat.setResellerID(Integer.valueOf(String.valueOf(objs[1])));
-					beat.setName(String.valueOf(objs[2]));
-					beat.setDescription(String.valueOf(objs[3]));
-					beat.setCoverageSchedule(String.valueOf(objs[4]));
-					beat.setDistance(Integer.valueOf(String.valueOf(objs[5])));
-				}
-			}
+			beat = getBeatForArea(session, areaID, tenantID);
 			
 			if(area != null && beat != null){
 				area.setBeat(beat);
@@ -114,7 +126,7 @@ public class AreaDAOImpl implements AreaDAO {
 	}
 
 	@Override
-	public void delete(int areaID) throws Exception{
+	public void delete(int areaID, int tenantID) throws Exception{
 		Session session = null;
 		Transaction transaction = null;
 		try{
@@ -122,8 +134,9 @@ public class AreaDAOImpl implements AreaDAO {
 			Area area = (Area)session.get(Area.class, areaID);
 			transaction = session.beginTransaction();
 			//Remove from BEAT_AREA
-			SQLQuery beatAreaQuery = session.createSQLQuery("DELETE FROM BEAT_AREA WHERE AREA_ID= ?");
+			SQLQuery beatAreaQuery = session.createSQLQuery("DELETE FROM BEAT_AREA WHERE AREA_ID= ? and TENANT_ID = ?");
 			beatAreaQuery.setParameter(0, areaID);
+			beatAreaQuery.setParameter(1, areaID);
 			beatAreaQuery.executeUpdate();
 			//Delete Area
 			session.delete(area);
@@ -143,16 +156,19 @@ public class AreaDAOImpl implements AreaDAO {
 	}
 
 	@Override
-	public List<Area> getResellerAreas(int resellerID) {
+	public List<Area> getTenantAreas(int tenantID) {
 		Session session = null;
 		List<Area> areas = null; 
 		try{
 			session = sessionFactory.openSession();
-			Query query = session.createQuery("from Area where resellerID = :resellerID order by DATE_CREATED DESC");
-			query.setParameter("resellerID", resellerID);
+			Query query = session.createQuery("from Area where tenantID = :tenantID order by DATE_CREATED DESC");
+			query.setParameter("tenantID", tenantID);
 			areas = query.list();
+			for(Area area : areas) {
+				area.setBeat(getBeatForArea(session, area.getAreaID(), tenantID));
+			}
 		}catch(Exception exception){
-			logger.error("Error while getting reseller Areas", exception);
+			logger.error("Error while getting tenant Areas", exception);
 		}finally{
 			if(session != null){
 				session.close();
@@ -162,25 +178,27 @@ public class AreaDAOImpl implements AreaDAO {
 	}
 
 	@Override
-	public List<Area> getBeatAreas(int beatID) {
+	public List<Area> getBeatAreas(int beatID, int tenantID) {
 
 		Session session = null;
 		List<Area> areaList = new ArrayList<Area>();
 		try {
 			session = sessionFactory.openSession();
 			//Get Areas
-			SQLQuery areasQuery = session.createSQLQuery("SELECT b.BEAT_ID, a.* FROM AREA A, BEAT_AREA b WHERE a.ID = b.AREA_ID AND b.BEAT_ID= ?");
+			SQLQuery areasQuery = session.createSQLQuery("SELECT b.BEAT_ID, a.* FROM AREA a, BEAT_AREA b WHERE a.ID = b.AREA_ID AND a.TENANT_ID = b.TENANT_ID AND b.BEAT_ID= ? AND a.TENANT_ID = ?");
 			areasQuery.setParameter(0, beatID);
+			areasQuery.setParameter(1, tenantID);
 			List areas = areasQuery.list();
 			for (Object obj : areas) {
 				Object[] objs = (Object[]) obj;
 				Area area = new Area();
 				area.setAreaID(Integer.valueOf(String.valueOf(objs[1])));
-				area.setResellerID(Integer.valueOf(String.valueOf(objs[2])));
+				area.setCode(String.valueOf(objs[2]));
 				area.setName(String.valueOf(objs[3]));
 				area.setDescription(String.valueOf(objs[4]));
 				area.setWordNo(String.valueOf(objs[5]));
 				area.setPinCode(String.valueOf(objs[6]));
+				area.setTenantID(Integer.valueOf(String.valueOf(objs[7])));
 				areaList.add(area);
 			}
 		} catch (Exception exception) {
@@ -195,25 +213,26 @@ public class AreaDAOImpl implements AreaDAO {
 	}
 	
 	@Override
-	public List<Area> getResellerAreasNotMappedToBeat(int resellerID) {
+	public List<Area> getTenantAreasNotMappedToBeat(int tenantID) {
 
 		Session session = null;
 		List<Area> areaList = new ArrayList<Area>();
 		try {
 			session = sessionFactory.openSession();
 			//Get Areas
-			SQLQuery areasQuery = session.createSQLQuery("SELECT * FROM AREA WHERE ID NOT IN (SELECT AREA_ID FROM BEAT_AREA) AND RESELLER_ID= ? ");
-			areasQuery.setParameter(0, resellerID);
+			SQLQuery areasQuery = session.createSQLQuery("SELECT * FROM AREA WHERE ID NOT IN (SELECT AREA_ID FROM BEAT_AREA) AND TENANT_ID= ? ");
+			areasQuery.setParameter(0, tenantID);
 			List areas = areasQuery.list();
 			for (Object obj : areas) {
 				Object[] objs = (Object[]) obj;
 				Area area = new Area();
 				area.setAreaID(Integer.valueOf(String.valueOf(objs[0])));
-				area.setResellerID(Integer.valueOf(String.valueOf(objs[1])));
+				area.setCode(String.valueOf(objs[1]));
 				area.setName(String.valueOf(objs[2]));
 				area.setDescription(String.valueOf(objs[3]));
 				area.setWordNo(String.valueOf(objs[4]));
 				area.setPinCode(String.valueOf(objs[5]));
+				area.setTenantID(Integer.valueOf(String.valueOf(objs[6])));
 				areaList.add(area);
 			}
 		} catch (Exception exception) {
@@ -228,26 +247,27 @@ public class AreaDAOImpl implements AreaDAO {
 	}
 	
 	@Override
-	public List<Area> getResellerAreasNotMappedToBeatForEdit(int resellerID, int beatID) {
+	public List<Area> getTenantAreasNotMappedToBeatForEdit(int tenantID, int beatID) {
 
 		Session session = null;
 		List<Area> areaList = new ArrayList<Area>();
 		try {
 			session = sessionFactory.openSession();
 			//Get Areas
-			SQLQuery areasQuery = session.createSQLQuery("SELECT * FROM AREA WHERE ID NOT IN (SELECT AREA_ID FROM BEAT_AREA) AND RESELLER_ID= ? UNION SELECT * FROM AREA WHERE ID IN (SELECT AREA_ID FROM BEAT_AREA WHERE BEAT_ID= ?)");
-			areasQuery.setParameter(0, resellerID);
+			SQLQuery areasQuery = session.createSQLQuery("SELECT * FROM AREA WHERE ID NOT IN (SELECT AREA_ID FROM BEAT_AREA) AND TENANT_ID= ? UNION SELECT * FROM AREA WHERE ID IN (SELECT AREA_ID FROM BEAT_AREA WHERE BEAT_ID= ?)");
+			areasQuery.setParameter(0, tenantID);
 			areasQuery.setParameter(1, beatID);
 			List areas = areasQuery.list();
 			for (Object obj : areas) {
 				Object[] objs = (Object[]) obj;
 				Area area = new Area();
 				area.setAreaID(Integer.valueOf(String.valueOf(objs[0])));
-				area.setResellerID(Integer.valueOf(String.valueOf(objs[1])));
+				area.setCode(String.valueOf(objs[1]));
 				area.setName(String.valueOf(objs[2]));
 				area.setDescription(String.valueOf(objs[3]));
 				area.setWordNo(String.valueOf(objs[4]));
 				area.setPinCode(String.valueOf(objs[5]));
+				area.setTenantID(Integer.valueOf(String.valueOf(objs[6])));
 				areaList.add(area);
 			}
 		} catch (Exception exception) {
